moved local planning functions and logic out of unitree_go2.py
